4.0.0 Stability - [4644] allow collapsing page builder tools panel - Related UI layout changes
