+++ ui/app/src/components/SiteSwitcherSelect/SiteSwitcherSelect.tsx
@@ -1,0 +1,45 @@
+/*
+ * Copyright (C) 2007-2021 Crafter Software Corporation. All Rights Reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as published by
+ * the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import * as React from 'react';
+import MenuItem from '@material-ui/core/MenuItem';
+import { FormattedMessage } from 'react-intl';
+import Select, { SelectProps } from '@material-ui/core/Select';
+
+export interface SiteSwitcherSelectProps extends SelectProps {
+  sites: { id: string; name: string }[];
+  classes?: SelectProps['classes'] & Partial<Record<'menuItem', string>>;
+}
+
+function SiteSwitcherSelect(props: SiteSwitcherSelectProps) {
+  const { sites, classes: { menuItem, ...classes } = {} } = props;
+  return (
+    <Select {...props} classes={classes}>
+      {sites.length === 0 && (
+        <MenuItem value="">
+          <FormattedMessage id="siteSwitcherSelected.siteSelectorNoSiteSelected" defaultMessage="Choose site" />
+        </MenuItem>
+      )}
+      {sites.map(({ id, name }) => (
+        <MenuItem key={id} value={id} className={menuItem}>
+          {name}
+        </MenuItem>
+      ))}
+    </Select>
+  );
+}
+
+export default SiteSwitcherSelect;
