+++ ui/app/src/modules/Preview/ToolBar.tsx
@@ -14,14 +14,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, { useEffect, useState } from 'react';
-import IconButton from '@material-ui/core/IconButton';
+import React from 'react';
 import ViewToolbar from '../../components/ViewToolbar';
-import Select from '@material-ui/core/Select';
-import MenuItem from '@material-ui/core/MenuItem';
-import { createStyles, makeStyles, Theme, useTheme } from '@material-ui/core/styles';
-import Paper from '@material-ui/core/Paper';
-import RefreshRounded from '@material-ui/icons/RefreshRounded';
 import LauncherOpenerButton from '../../components/LauncherOpenerButton';
 import {
   changeCurrentUrl,
@@ -41,29 +35,19 @@ import {
   useSiteList
 } from '../../utils/hooks';
 import { getHostToGuestBus } from './previewContext';
-import { isBlank } from '../../utils/string';
-import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
+import { defineMessages, useIntl } from 'react-intl';
 import QuickCreate from './QuickCreate';
 import { changeSite } from '../../state/reducers/sites';
-import Switch from '@material-ui/core/Switch';
 import Tooltip from '@material-ui/core/Tooltip';
-import withStyles from '@material-ui/core/styles/withStyles';
-import palette from '../../styles/palette';
-import SingleItemSelector from '../Content/Authoring/SingleItemSelector';
-import { DetailedItem } from '../../models/Item';
-import PagesSearchAhead from '../../components/PagesSearchAhead/PagesSearchAhead';
-import clsx from 'clsx';
-import { AllItemActions, generateSingleItemOptions, itemActionDispatcher } from '../../utils/itemActions';
-import ActionsGroup from '../../components/ActionsGroup';
-import Skeleton from '@material-ui/lab/Skeleton';
-import useMediaQuery from '@material-ui/core/useMediaQuery';
 import { setSiteCookie } from '../../utils/auth';
 import LogoAndMenuBundleButton from '../../components/LogoAndMenuBundleButton';
 import { getSystemLink } from '../../components/LauncherSection';
 import { hasCreateAction, hasEditAction } from '../../utils/content';
-import ItemDisplay from '../../components/ItemDisplay';
-import Typography from '@material-ui/core/Typography';
 import { PublishingStatusButton } from '../../components/PublishingStatusButton';
+import EditModeSwitch from '../../components/EditModeSwitch';
+import { AddressBar } from '../../components/PreviewAddressBar/PreviewAddressBar';
+import SiteSwitcherSelect, { useSiteSwitcherMinimalistStyles } from '../../components/SiteSwitcherSelect';
+import { isBlank } from '../../utils/string';
 
 const translations = defineMessages({
   openToolsPanel: {
@@ -78,241 +62,11 @@ const translations = defineMessages({
     id: 'common.toggleSidebarTooltip',
     defaultMessage: 'Toggle sidebar'
   },
-  reload: {
-    id: 'words.reload',
-    defaultMessage: 'Reload'
-  },
   itemMenu: {
     id: 'previewToolbar.itemMenu',
     defaultMessage: 'Item menu'
   }
 });
-
-const foo = () => void 0;
-
-export const EditSwitch = withStyles({
-  checked: {
-    color: palette.green.tint
-  }
-})(Switch);
-
-const useAddressBarStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    toolbar: {
-      placeContent: 'center space-between'
-    },
-    addressBarInput: {
-      width: 400,
-      padding: '2px 4px',
-      display: 'flex',
-      alignItems: 'center',
-      backgroundColor: theme.palette.background.default
-    },
-    addressBarInputFocused: {
-      border: `1px solid ${theme.palette.primary.main}`,
-      backgroundColor: theme.palette.background.paper
-    },
-    inputContainer: {
-      marginLeft: theme.spacing(1)
-    },
-    input: {
-      border: 'none',
-      background: 'transparent',
-      '&:focus:invalid, &:focus': {
-        border: 'none',
-        boxShadow: 'none'
-      }
-    },
-    siteSwitcherSelectMenu: {
-      maxWidth: 110,
-      padding: '10px 10px'
-    },
-    siteSwitcherSelectMenuRoot: {
-      '&.MuiInput-underline::before': {
-        display: 'none'
-      },
-      '&.MuiInput-underline::after': {
-        display: 'none'
-      },
-      background: 'transparent'
-    },
-    siteSwitcherMenuItem: {
-      maxWidth: 390,
-      whiteSpace: 'nowrap',
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      display: 'block'
-    },
-    divider: {
-      height: 28,
-      margin: 4
-    },
-    selectorPopoverRoot: {
-      width: 400,
-      marginLeft: '4px'
-    },
-    hidden: {
-      visibility: 'hidden'
-    },
-    itemActionSkeleton: {
-      width: 40,
-      margin: '0 5px'
-    },
-    itemDisplayWrapper: {
-      width: '100%',
-      overflow: 'hidden',
-      cursor: 'pointer',
-      display: 'flex'
-    },
-    itemPreviewUrl: {
-      overflow: 'hidden',
-      whiteSpace: 'nowrap',
-      textOverflow: 'ellipsis',
-      marginLeft: '4px'
-    },
-    itemDisplaySkeleton: {
-      marginLeft: '5px',
-      width: '100%'
-    }
-  })
-);
-
-interface AddressBarProps {
-  site: string;
-  url: string;
-  item?: DetailedItem;
-  onSiteChange: (siteId: string) => any;
-  onUrlChange: (value: string) => any;
-  onRefresh: (e) => any;
-  sites: { id: string; name: string }[];
-}
-
-export function AddressBar(props: AddressBarProps) {
-  const classes = useAddressBarStyles();
-  const { formatMessage } = useIntl();
-  const { site, url = '', sites = [], item, onSiteChange = foo, onUrlChange = foo, onRefresh = foo } = props;
-  const noSiteSet = isBlank(site);
-  const [internalUrl, setInternalUrl] = useState(url);
-  const [openSelector, setOpenSelector] = useState(false);
-  const [focus, setFocus] = useState(false);
-  const disabled = noSiteSet || !item;
-
-  useEffect(() => {
-    url && setInternalUrl(url);
-  }, [url]);
-
-  const theme = useTheme();
-  const [numOfVisibleActions, setNumOfVisibleActions] = useState(5);
-  const isSmallScreen = useMediaQuery(theme.breakpoints.only('sm'));
-  const isMediumScreen = useMediaQuery(theme.breakpoints.only('md'));
-  const isLargeScreen = useMediaQuery(theme.breakpoints.only('lg'));
-  useEffect(() => {
-    setNumOfVisibleActions(isSmallScreen ? 1 : isMediumScreen ? 4 : isLargeScreen ? 8 : 15);
-  }, [isSmallScreen, isMediumScreen, isLargeScreen]);
-
-  const onSiteChangeInternal = (value) => !isBlank(value) && value !== site && onSiteChange(value);
-
-  const { authoringBase } = useEnv();
-  const dispatch = useDispatch();
-  const clipboard = useSelection((state) => state.content.clipboard);
-  const onMenuItemClicked = (option: AllItemActions, event: React.MouseEvent<HTMLLIElement, MouseEvent>) =>
-    itemActionDispatcher({
-      site,
-      item,
-      option,
-      authoringBase,
-      dispatch,
-      formatMessage,
-      clipboard,
-      event
-    });
-  const actions = generateSingleItemOptions(item, formatMessage)?.flatMap((options) => options);
-
-  return (
-    <>
-      <IconButton title={formatMessage(translations.reload)} onClick={onRefresh}>
-        <RefreshRounded />
-      </IconButton>
-      <Paper
-        variant={focus ? 'elevation' : 'outlined'}
-        elevation={focus ? 2 : 0}
-        className={clsx(classes.addressBarInput, focus && classes.addressBarInputFocused)}
-      >
-        <Select
-          value={site}
-          className={classes.siteSwitcherSelectMenuRoot}
-          classes={{
-            select: classes.input,
-            selectMenu: classes.siteSwitcherSelectMenu
-          }}
-          onChange={({ target: { value } }) => onSiteChangeInternal(value)}
-          displayEmpty
-          variant="standard"
-        >
-          {noSiteSet && (
-            <MenuItem value="">
-              <FormattedMessage id="previewToolBar.siteSelectorNoSiteSelected" defaultMessage="Choose site" />
-            </MenuItem>
-          )}
-          {sites.map(({ id, name }) => (
-            <MenuItem key={id} value={id} className={classes.siteSwitcherMenuItem}>
-              {name}
-            </MenuItem>
-          ))}
-        </Select>
-        {!focus && item && (
-          <div className={classes.itemDisplayWrapper} onClick={() => setFocus(true)}>
-            <ItemDisplay item={item} styles={{ root: { maxWidth: '100%' } }} />
-            <Typography className={classes.itemPreviewUrl} color="textSecondary">
-              • {item.previewUrl}
-            </Typography>
-          </div>
-        )}
-        {(focus || !item) && (
-          <PagesSearchAhead
-            value={internalUrl}
-            placeholder={noSiteSet ? '' : '/'}
-            disabled={disabled}
-            onEnter={(value) => onUrlChange(value)}
-            classes={{
-              input: classes.input
-            }}
-            onFocus={() => setFocus(true)}
-            onBlur={() => setFocus(false)}
-          />
-        )}
-        <SingleItemSelector
-          disabled={disabled}
-          rootPath="/site/website/index.xml"
-          selectedItem={item as DetailedItem}
-          open={openSelector}
-          onClose={() => setOpenSelector(false)}
-          onDropdownClick={() => setOpenSelector(!openSelector)}
-          onItemClicked={(item) => {
-            setOpenSelector(false);
-            setInternalUrl(item.previewUrl);
-            onUrlChange(item.previewUrl);
-          }}
-          hideUI
-          classes={{
-            popoverRoot: classes.selectorPopoverRoot
-          }}
-        />
-      </Paper>
-      {item ? (
-        <ActionsGroup max={numOfVisibleActions} actions={actions} onActionClicked={onMenuItemClicked} />
-      ) : (
-        <>
-          <Skeleton animation="pulse" className={classes.itemActionSkeleton} />
-          <Skeleton animation="pulse" className={classes.itemActionSkeleton} />
-          <Skeleton animation="pulse" className={classes.itemActionSkeleton} />
-          <Skeleton animation="pulse" className={classes.itemActionSkeleton} />
-          <Skeleton animation="pulse" variant="circle" width="25px" height="25px" />
-        </>
-      )}
-    </>
-  );
-}
 
 export default function ToolBar() {
   const { formatMessage } = useIntl();
@@ -330,6 +84,26 @@ export default function ToolBar() {
   const { authoringBase } = useEnv();
   const write = hasEditAction(item?.availableActions);
   const createContent = hasCreateAction(item?.availableActions);
+  const classes = useSiteSwitcherMinimalistStyles();
+
+  const onSiteChange = ({ target: { value } }) => {
+    if (!isBlank(value) && site !== value) {
+      if (previewChoice[value] === '2') {
+        dispatch(changeSite(value));
+      } else {
+        setSiteCookie(value);
+        setTimeout(
+          () =>
+            (window.location.href = getSystemLink({
+              site: value,
+              systemLinkId: 'preview',
+              previewChoice,
+              authoringBase
+            }))
+        );
+      }
+    }
+  };
 
   return (
     <ViewToolbar>
@@ -340,9 +114,34 @@ export default function ToolBar() {
             onClick={() => dispatch(showToolsPanel ? closeTools() : openTools())}
           />
         </Tooltip>
+        <SiteSwitcherSelect
+          value={site}
+          sites={sites}
+          displayEmpty
+          variant="standard"
+          className={classes.menuRoot}
+          style={{ marginRight: 5 }}
+          onChange={onSiteChange}
+          classes={{
+            select: classes.input,
+            selectMenu: classes.menu,
+            menuItem: classes.menuItem
+          }}
+        />
         <QuickCreate disabled={!createContent} />
+      </section>
+      <section>
+        <AddressBar
+          site={site ?? ''}
+          url={computedUrl}
+          item={item}
+          onUrlChange={(url) => dispatch(changeCurrentUrl(url))}
+          onRefresh={() => getHostToGuestBus().next({ type: RELOAD_REQUEST })}
+        />
+      </section>
+      <section>
         <Tooltip title={!write ? '' : formatMessage(translations.toggleEditMode)}>
-          <EditSwitch
+          <EditModeSwitch
             disabled={!write}
             color="default"
             checked={editMode}
@@ -351,32 +150,6 @@ export default function ToolBar() {
             }}
           />
         </Tooltip>
-        <AddressBar
-          site={site ?? ''}
-          sites={sites}
-          url={computedUrl}
-          item={item}
-          onSiteChange={(site) => {
-            if (previewChoice[site] === '2') {
-              dispatch(changeSite(site));
-            } else {
-              setSiteCookie(site);
-              setTimeout(
-                () =>
-                  (window.location.href = getSystemLink({
-                    site,
-                    systemLinkId: 'preview',
-                    previewChoice,
-                    authoringBase
-                  }))
-              );
-            }
-          }}
-          onUrlChange={(url) => dispatch(changeCurrentUrl(url))}
-          onRefresh={() => getHostToGuestBus().next({ type: RELOAD_REQUEST })}
-        />
-      </section>
-      <section>
         <PublishingStatusButton variant="icon" />
         <LauncherOpenerButton sitesRailPosition="left" icon="apps" />
       </section>
